拆分test模块，在Lite上去掉了common-io/json4s/commons-lang3/commons-text/emoji依赖
