+++ duckling-fork-chinese/version.sbt
@@ -1,1 +1,1 @@
-ThisBuild / version := "1.2-SNAPSHOT"
+ThisBuild / version := "1.3-SNAPSHOT"
